Naively try leveraging the tests+coverage script from centralized actions
On the last commit we had trouble that the local condamerge script can't handle more than two env files. This doesn't bother fixing that, but sidesteps it to try a different job based on a centralized script. In another PR QE is not playing nicely with all the tests together, so let's see if we can get as far as reproducing that.